[FIX] (Raise error when inexistent id)

+++ api/v1/views/states.py
@@ -31,6 +31,8 @@ def delete_states(state_id):
     """ Deletes a state """
     try:
         state = storage.get(State, state_id)
+        if state is None:
+            abort(404)
         storage.delete(state)
         storage.save()
         return jsonify({}), 200
